project/backend/requirements.txt
feedparser==6.0.8
fastapi
uvicorn
PyYAML
python-dateutil